Since r5086 didn't break any builds, reap the benefit of tgmath.
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@5097 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Cocoa/OOCASound.m
@@ -361,9 +361,9 @@ static BOOL VerifyOneBuffer(AudioBuffer *buffer, OOUInteger numFrames, float *ba
 		float val = floatBuffer[i];
 		BOOL abnormal = isnan(val) || !isfinite(val);
 		
-		if (abnormal || (fabsf(val) > kOOAudioSlop))
+		if (abnormal || (fabs(val) > kOOAudioSlop))
 		{
-			if (abnormal || (!abnormal && !worstAbnormal && fabsf(*badVal < fabsf(val))))
+			if (abnormal || (!abnormal && !worstAbnormal && fabs(*badVal < fabs(val))))
 			{
 				worstAbnormal = abnormal;
 				*badVal = val;
