+++ src/Core/OOPlanetDrawable.m
@@ -158,7 +158,7 @@
 
 - (void) setRadius:(float)radius
 {
-	_radius = fabsf(radius);
+	_radius = fabs(radius);
 	[self recalculateTransform];
 }
 
@@ -171,14 +171,14 @@
 
 - (void) setLevelOfDetail:(float)lod
 {
-	_lod = roundf(OOClamp_0_1_f(lod) * LOD_GRANULARITY);
+	_lod = round(OOClamp_0_1_f(lod) * LOD_GRANULARITY);
 }
 
 
 - (void) calculateLevelOfDetailForViewDistance:(float)distance
 {
 	float	drawFactor = [[UNIVERSE gameView] viewSize].width / 100.0f;
-	float	drawRatio2 = drawFactor * _radius / sqrtf(distance); // proportional to size on screen in pixels
+	float	drawRatio2 = drawFactor * _radius / sqrt(distance); // proportional to size on screen in pixels
 	
 	float lod = sqrt(drawRatio2 * LOD_FACTOR);
 	if ([UNIVERSE reducedDetail])
